+++ src/Managers/LightingManager.cpp
@@ -16,6 +16,7 @@ LightingManager::~LightingManager()
 
 bool LightingManager::Initialize()
 {
+    // this HAS to be suboptimal in some way
     // this will be sorted out later
     // atm, you can only switch between the different light types by changing the type in the editor
     // this is NOT the way i wanted to do it originally but will keep it like this until i create a separate gui system
@@ -43,41 +44,72 @@ void LightingManager::Update()
     lightData.ambientColor = DirectX::XMFLOAT4(0.1f, 0.1f, 0.1f, 1.0f);
     lightData.lightType = 0;
 
-    auto lightEntities = registry.GetEntitiesWith<LightComponent>();
+    // now we can just go through each light type entity separately
 
-    if (!lightEntities.empty())
+    auto directionalLights = registry.GetEntitiesWith<DirectionalLightComponent>();
+    if (!directionalLights.empty())
     {
-        auto entity = *lightEntities.begin();
-        auto *light = registry.GetComponent<LightComponent>(entity);
-        auto *transform = registry.GetComponent<TransformComponent>(entity);
+        auto entity = *directionalLights.begin();
+        auto *light = registry.GetComponent<DirectionalLightComponent>(entity);
 
-        if (light && transform)
+        if (light)
         {
             lightData.ambientColor = light->ambientColor;
             lightData.diffuseColor = light->diffuseColor;
             lightData.lightIntensity = light->intensity;
 
-            lightData.lightType = static_cast<int>(light->type);
+            // everything after this is for that specific light type
+            // obviously we could type 0 here but like ehhhh
+            lightData.lightType = static_cast<int>(light->GetLightType());
+            lightData.lightDirection = light->direction;
+        }
+    }
 
-            switch (light->type)
-            {
-            case LightComponent::LightType::Directional:
-                lightData.lightDirection = light->direction;
-                break;
+    // overrides directional light if there is a point light
 
-            case LightComponent::LightType::Point:
-                lightData.lightPosition = transform->position;
-                lightData.lightRange = light->range;
-                break;
+    auto pointLights = registry.GetEntitiesWith<PointLightComponent>();
+    if (!pointLights.empty())
+    {
+        auto entity = *pointLights.begin();
+        auto *light = registry.GetComponent<PointLightComponent>(entity);
 
-            case LightComponent::LightType::Spot:
-                lightData.lightPosition = transform->position;
-                lightData.lightDirection = light->direction;
-                lightData.lightRange = light->range;
-                lightData.spotInnerCone = cosf(DirectX::XMConvertToRadians(30.0f));
-                lightData.spotOuterCone = cosf(DirectX::XMConvertToRadians(45.0f));
-                break;
-            }
+        if (light)
+        {
+            lightData.ambientColor = light->ambientColor;
+            lightData.diffuseColor = light->diffuseColor;
+            lightData.lightIntensity = light->intensity;
+
+            // everything after this is for that specific light type
+            // obviously we could type 1 here but like ehhhh
+            lightData.lightType = static_cast<int>(light->GetLightType());
+            lightData.lightPosition = light->position;
+            lightData.lightRange = light->range;
+        }
+    }
+
+    // overrides everything else since it's last
+
+    auto spotLights = registry.GetEntitiesWith<SpotLightComponent>();
+    if (!spotLights.empty())
+    {
+        auto entity = *spotLights.begin();
+        auto *light = registry.GetComponent<SpotLightComponent>(entity);
+
+        if (light)
+        {
+            lightData.ambientColor = light->ambientColor;
+            lightData.diffuseColor = light->diffuseColor;
+            lightData.lightIntensity = light->intensity;
+
+            // everything after this is for that specific light type
+            // obviously we could type 2 here but like ehhhh
+            lightData.lightType = static_cast<int>(light->GetLightType());
+            lightData.lightPosition = light->position;
+            lightData.lightDirection = light->direction;
+            lightData.lightRange = light->range;
+            // convert to radians maybe?
+            lightData.spotInnerCone = cosf(light->innerConeAngle);
+            lightData.spotOuterCone = cosf(light->outerConeAngle);
         }
     }